requirements.txt
# --------- pytorch --------- #
torch==2.0.1
torchvision==0.15.2
lightning==2.0.0
torchmetrics==0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
tensorboard

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
ipykernel
scikit-image
matplotlib
